pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "csrsb"
version = "0.1.0"
description = "Claude Screen Recorder Skill Builder — record a workflow, generate a Claude Code Skill."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Claude Screen Recorder Skill Builder contributors" }]
dependencies = [
    "anthropic>=0.92.0",
    "pydantic>=2.6",
    "click>=8.1",
    "jinja2>=3.1",
    "pynput>=1.7.6",
    "mss>=9.0",
    "pillow>=10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
]

[project.scripts]
csrsb = "csrsb.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
csrsb = ["translator/templates/*.j2"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
